+++ src/level3/DivideOperator.java
@@ -2,7 +2,7 @@ package level3;
 
 public class DivideOperator implements Operator {
     @Override
-    public double operate(int firestNum, int secondNum) {
+    public double operate(double firestNum, double secondNum) {
         return (double) firestNum / secondNum;
     }
 }
