+++ src/level3/ModOperator.java
@@ -2,7 +2,7 @@ package level3;
 
 public class ModOperator implements Operator {
     @Override
-    public double operate(int firstNum, int secondNum) {
+    public double operate(double firstNum, double secondNum) {
         return firstNum % secondNum;
     }
 }
