+++ src/level3/Operator.java
@@ -1,5 +1,5 @@
 package level3;
 
 public interface Operator {
-    double operate(int firstNum, int secondNum);
+    double operate(double firstNum, double secondNum);
 }
