+++ src/level3/SubstractOperator.java
@@ -2,7 +2,7 @@ package level3;
 
 public class SubstractOperator implements Operator {
     @Override
-    public double operate(int firstNum, int secondNum){
+    public double operate(double firstNum, double secondNum){
         return firstNum - secondNum;
     }
 }
